+++ cmd/operator-sdk/scorecard/cmd.go
@@ -37,7 +37,9 @@ func NewCmd() *cobra.Command {
 	scorecardCmd.Flags().String(scorecard.ConfigOpt, "", fmt.Sprintf("config file (default is '<project_dir>/%s'; the config file's extension and format can be .yaml, .json, or .toml)", scorecard.DefaultConfigFile))
 	scorecardCmd.Flags().String(scplugins.KubeconfigOpt, "", "Path to kubeconfig of custom resource created in cluster")
 	scorecardCmd.Flags().StringP(scorecard.OutputFormatOpt, "o", scorecard.TextOutputFormat, fmt.Sprintf("Output format for results. Valid values: %s, %s", scorecard.TextOutputFormat, scorecard.JSONOutputFormat))
-	scorecardCmd.Flags().String(schelpers.VersionOpt, schelpers.DefaultScorecardVersion, fmt.Sprintf("scorecard version (tech preview version is '%s'", schelpers.LatestScorecardVersion))
+	scorecardCmd.Flags().String(schelpers.VersionOpt, schelpers.DefaultScorecardVersion, "scorecard version. Valid values: v1alpha1, v1alpha2")
+	scorecardCmd.Flags().StringP(schelpers.SelectorOpt, "l", scplugins.DefaultSelector, "selector (label query) to filter tests on (only valid when version is v1alpha2)")
+	scorecardCmd.Flags().Bool(schelpers.DryRunOpt, false, "If true, only print the tests that would be run.")
 
 	// TODO: make config file global and make this a top level flag
 	viper.BindPFlag(scorecard.ConfigOpt, scorecardCmd.Flags().Lookup(scorecard.ConfigOpt))
@@ -45,6 +47,8 @@ func NewCmd() *cobra.Command {
 	viper.BindPFlag("scorecard."+scplugins.KubeconfigOpt, scorecardCmd.Flags().Lookup(scplugins.KubeconfigOpt))
 	viper.BindPFlag("scorecard."+scorecard.OutputFormatOpt, scorecardCmd.Flags().Lookup(scorecard.OutputFormatOpt))
 	viper.BindPFlag("scorecard."+schelpers.VersionOpt, scorecardCmd.Flags().Lookup(schelpers.VersionOpt))
+	viper.BindPFlag("scorecard."+schelpers.SelectorOpt, scorecardCmd.Flags().Lookup(schelpers.SelectorOpt))
+	viper.BindPFlag("scorecard."+schelpers.DryRunOpt, scorecardCmd.Flags().Lookup(schelpers.DryRunOpt))
 
 	return scorecardCmd
 }